requirements.txt
celery<5
django<3
django-admin-multiple-choice-list-filter
django-admin-tools
django-autocomplete-light
django-compressor
django-debug-toolbar==3.2.4
django-import-export==2.7.1
django-redis
djangorestframework==3.13.1
flower==0.9.7
openpyxl
requests
pandas
pillow
psycopg2-binary
six
xlrd
xlwt
xlsxwriter
git+https://rown@bitbucket.org/rown/django-cache-model.git
git+https://rown@bitbucket.org/rown/django-preferences.git
git+https://rown@bitbucket.org/rown/django-public-model.git
git+https://rown@bitbucket.org/rown/django-sort-model.git
pytest
pytest-django
pytest-cov
